mediamtx_communicator/requirements.txt
fastapi
uvicorn
kafka-python
python-multipart
requests
httpx